backend/requirements.txt
asgiref==3.6.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
fonttools==4.39.2
install==1.3.5
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
matplotlib==3.7.1
nltk==3.8.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
psycopg2==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
regex==2022.10.31
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post1
sqlparse==0.4.3
threadpoolctl==3.1.0
tqdm==4.65.0
